Mastermind initial commit.
I have the Elm Architecture working for NCurses.
The game is "playable" - use "End" to quit, and color keys to play.  Currently much of View doesn't work, and Update has some issues too - with result calculation and figuring out if you have won or not.  I'm pushing this up for Dave to look at the TEA framework, though.